feat: working bridge between go and zig.
Use zig to expose a C ABI that will initialize TB contexts which will
eventually contain the zig client. The context is then passed to other
exposed functions. The C ABI takes in function pointers to callbacks.

cgo is used to include the C header so that Go can instantiate a TB
context. The exposed C `init` function is then called to initialize it.
Go then passes callbacks to the C ABI as pointers.