Go client
Go client definition that uses the C ABI to `init` a TigerBeetle
context. This exposes a `Request` method that takes in raw bytes and
user callbacks to allow users to create their own wrappers as they see
fit.

The user callback and contexts are stored in a global registry as Go
pointers are not allowed to be passed to C. The index in the registry
is passed to C instead for retrieval when results are returned. A mutex
is used to provide thread safety when accessing the registry.

Convenience wrappers for `CreateAccounts`, `LookupAccounts`,
`CreateTransfers`, `CommitTransfers` and `LookupTransfers` that uses
channels for synchronization are provided.

Basic tests exist to ensure functionality.